create outpack server client
