+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/OutpackServerAPI.kt
@@ -1,0 +1,9 @@
+package org.vaccineimpact.orderlyweb
+
+import okhttp3.OkHttpClient
+import org.vaccineimpact.orderlyweb.db.Config
+
+class OutpackServer(
+        config: Config,
+        client: OkHttpClient = OkHttpClient()
+) : PorcelainAPIServer("Outpack server", config["outpack.server"], client)
